add entrypoint script and update to court detection
